Tests for China UnionPay added
This code is very buggy. I've been looking, but can't figure out how to fix
the bugs. The tests for Maestro only works if I end my test with
.to.equal('China UnionPay') instead of .to.equal('Maestro') and when I enter
nextStep(); in the console it tells me I didn't achieve full test coverage
for the Maestro tests that show as passed in the browser. Unfortunately I
haven't been able to figure this out and cant move forward to the other tests
because of it. But if my code for Maestro was correct, my code for China
UnionPay should be correct as well.